+++ src/design/plone/contenttypes/interfaces/__init__.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 """Module where all interfaces, events and exceptions live."""
 from redturtle.volto.interfaces import IRedturtleVoltoLayer
+from redturtle.bandi.interfaces.browserlayer import IRedturtleBandiLayer
 from zope.interface import Interface
 
 
-class IDesignPloneContenttypesLayer(IRedturtleVoltoLayer):
+class IDesignPloneContenttypesLayer(IRedturtleVoltoLayer, IRedturtleBandiLayer):
     """Marker interface that defines a browser layer."""