+++ src/design/plone/contenttypes/testing.py
@@ -16,7 +16,6 @@ import collective.volto.blocksfield
 import collective.volto.enhancedlinks
 import collective.z3cform.datagridfield
 import design.plone.contenttypes
-import eea.api.taxonomy
 import kitconcept.seo
 import plone.app.caching
 import plone.formwidget.geolocation
@@ -43,7 +42,6 @@ class DesignPloneContenttypesLayer(RedturtleVoltoLayer):
         )
         self.loadZCML(package=redturtle.bandi)
         self.loadZCML(package=kitconcept.seo)
-        self.loadZCML(package=eea.api.taxonomy)
         self.loadZCML(package=collective.taxonomy)
         self.loadZCML(package=collective.z3cform.datagridfield)
 
@@ -75,7 +73,6 @@ class DesignPloneContenttypesRestApiLayer(RedturtleVoltoRestApiLayer):
         self.loadZCML(package=collective.volto.enhancedlinks)
         self.loadZCML(package=design.plone.contenttypes, context=configurationContext)
         self.loadZCML(package=plone.formwidget.geolocation)
-        self.loadZCML(package=eea.api.taxonomy)
         self.loadZCML(package=collective.taxonomy)
         self.loadZCML(package=collective.z3cform.datagridfield)
         xmlconfig.file(