+++ src/design/plone/contenttypes/upgrades/to_730x.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from plone.app.upgrade.utils import installOrReinstallProduct
 from design.plone.contenttypes.events.common import SUBFOLDERS_MAPPING
 from design.plone.contenttypes.utils import create_default_blocks
 from plone import api
@@ -161,6 +162,17 @@ def to_7308(context):
 
 
 def to_7309(context):
+    logger.info("Uninstall eea.api.taxonomy")
+    ps = api.portal.get_tool(name="portal_setup")
+    ps.runAllImportStepsFromProfile(
+        "profile-design.plone.contenttypes:remove_eea_api_taxonomy"
+    )
+    ps.unsetLastVersionForProfile("eea.api.taxonomy:default")
+    logger.info("Install blocksfield")
+    installOrReinstallProduct(api.portal.get(), "collective.volto.blocksfield")
+
+
+def to_7310(context):
     logger.info("Enable kitconcept.seo behavior to File")
     portal_types = api.portal.get_tool(name="portal_types")
     behavior = "kitconcept.seo"